changed println to printf, remove casts and change some variables type to double

+++ src/main/java/lesson6/animals/Dog.java
@@ -5,7 +5,7 @@ import lesson6.Animal;
 public class Dog extends Animal {
     protected String type;
 
-    public Dog(String name, int runLimit, int swimLimit, float jumpLimit, String type) {
+    public Dog(String name, double runLimit, double swimLimit, double jumpLimit, String type) {
         super(name, runLimit, swimLimit, jumpLimit);
         this.type = type;
     }
